+++ src/intcode/operations/JumpIfTrueOperation.java
@@ -7,7 +7,7 @@ import intcode.operations.Params.ParamType;
 
 public class JumpIfTrueOperation extends AbstractOperation{
 	
-	Integer nextInstructionCounter = null;
+	Long nextInstructionCounter = null;
 	
 	public JumpIfTrueOperation() {
 		setParameters(Arrays.asList(ParamType.READ, ParamType.READ));
@@ -24,7 +24,7 @@ public class JumpIfTrueOperation extends AbstractOperation{
 	}
 	
 	@Override
-	protected Integer getNewInstructionCounter(int i) {
+	protected long getNewInstructionCounter(long i) {
 		if (nextInstructionCounter != null) {
 			return nextInstructionCounter;
 		}
